# packages/contracts/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Spott API
  version: 1.0.0
  description: Spott iOS/Web 单一契约。具备 Idempotency-Key 参数的副作用写入由调用方提供稳定幂等键； 媒体上传、完成、附件与举报均以调用方保留的 UUID 和服务端持久回执支持响应丢失恢复； 业务状态始终由服务端裁决。
  contact:
    name: Spott Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.spott.jp/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /registrations/{id}/feedback:
    get:
      operationId: getOwnEventFeedback
      summary: Get the current user's private post-event feedback state
      tags:
        - Feedback
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Authoritative feedback eligibility and current private feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnFeedbackState'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    post:
      operationId: submitEventFeedback
      summary: Submit or edit private structured post-event feedback
      tags:
        - Feedback
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackInput'
      responses:
        '201':
          description: Feedback pending moderation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackReceipt'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /events/{id}/feedback-summary:
    get:
      operationId: getEventFeedbackSummary
      summary: Get privacy-thresholded aggregate feedback tags
      tags:
        - Feedback
      security: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Aggregate-only summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackSummary'
        '404':
          $ref: '#/components/responses/Error'
  /events/{id}/feedback/private:
    get:
      operationId: getPrivateEventFeedback
      summary: Get private improvement suggestions as the event host
      tags:
        - Feedback
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Private feedback items
        '403':
          $ref: '#/components/responses/Error'
  /me/achievements:
    get:
      operationId: getMyAchievements
      summary: List achievements for the current user
      tags:
        - Profiles
      responses:
        '200':
          description: Achievement list
        '401':
          $ref: '#/components/responses/Error'
  /me/achievements/evaluate:
    post:
      operationId: evaluateMyAchievements
      summary: Evaluate achievement definitions against current metrics
      tags:
        - Profiles
      responses:
        '200':
          description: Newly awarded achievements
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OpsSession:
      type: apiKey
      in: cookie
      name: __Host-spott_ops
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        oneOf:
          - $ref: '#/components/schemas/UUID'
          - type: string
            maxLength: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  responses:
    Error:
      description: Stable business error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    FeedbackInput:
      type: object
      required:
        - attendanceRating
        - tags
      properties:
        attendanceRating:
          type: integer
          minimum: 1
          maximum: 5
        tags:
          type: array
          maxItems: 5
          items:
            type: string
            enum:
              - friendly
              - well_organized
              - clear_information
              - safe
              - would_join_again
        comment:
          type: string
          maxLength: 500
        visibility:
          type: string
          enum:
            - private
            - aggregate_only
          default: aggregate_only
    FeedbackSubmissionState:
      type: string
      enum:
        - not_submitted
        - edit_available
        - edit_limit_reached
        - window_closed
        - not_eligible
    OwnFeedback:
      type: object
      required:
        - id
        - attendanceRating
        - tags
        - comment
        - visibility
        - moderationState
        - editCount
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        attendanceRating:
          type: integer
          minimum: 1
          maximum: 5
        tags:
          type: array
          maxItems: 5
          items:
            type: string
            enum:
              - friendly
              - well_organized
              - clear_information
              - safe
              - would_join_again
        comment:
          type:
            - string
            - 'null'
          maxLength: 500
        visibility:
          type: string
          enum:
            - private
            - aggregate_only
        moderationState:
          type: string
          minLength: 1
        editCount:
          type: integer
          minimum: 0
          maximum: 1
        createdAt:
          type:
            - string
            - 'null'
          format: date-time
        updatedAt:
          type:
            - string
            - 'null'
          format: date-time
    OwnFeedbackState:
      type: object
      required:
        - registrationId
        - eventId
        - state
        - canSubmit
        - canEdit
        - windowClosesAt
        - feedback
      properties:
        registrationId:
          $ref: '#/components/schemas/UUID'
        eventId:
          $ref: '#/components/schemas/UUID'
        state:
          $ref: '#/components/schemas/FeedbackSubmissionState'
        canSubmit:
          type: boolean
        canEdit:
          type: boolean
        windowClosesAt:
          type:
            - string
            - 'null'
          format: date-time
        feedback:
          anyOf:
            - $ref: '#/components/schemas/OwnFeedback'
            - type: 'null'
    FeedbackReceipt:
      type: object
      required:
        - id
        - eventId
        - status
        - editCount
        - rewardPoints
        - createdAt
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        eventId:
          $ref: '#/components/schemas/UUID'
        status:
          type: string
          enum:
            - pending_moderation
        editCount:
          type: integer
          minimum: 0
          maximum: 1
        rewardPoints:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
    FeedbackSummary:
      type: object
      required:
        - sampleSize
        - minimumSampleSize
        - published
        - tags
      properties:
        sampleSize:
          type: integer
          minimum: 0
        minimumSampleSize:
          type: integer
          minimum: 1
        published:
          type: boolean
        tags:
          type: array
          items:
            type: object
            required:
              - tag
              - count
              - rate
            properties:
              tag:
                type: string
                enum:
                  - friendly
                  - well_organized
                  - clear_information
                  - safe
                  - would_join_again
              count:
                type: integer
                minimum: 0
              rate:
                type: number
                minimum: 0
                maximum: 1
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
            - retryable
            - fieldErrors
            - actions
            - meta
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            retryable:
              type: boolean
            fieldErrors:
              type: array
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                  message:
                    type: string
            actions:
              type: array
              items:
                type: object
                required:
                  - type
                  - label
                properties:
                  type:
                    type: string
                  label:
                    type: string
            meta:
              type: object
              additionalProperties: true